Improve install script
Make the script location-independent. That is, the repository need not
be checked out to ~/.dotfiles anymore.

Fix an issue where symlinks to directories would be properly created
during the first run of the script, but would create a symlink WITHIN
the existing symlink upon all successive runs.

Also make the install script less verbose.